Convert eslint-plugin-test-rules/index.js to Typescript
Fixes #576

Convert `eslint-plugin-test-rules/index.js` to TypeScript.

* Rename `eslint-plugin-test-rules/index.js` to `eslint-plugin-test-rules/index.ts`.
* Update `module.exports` to `export` and add TypeScript types for ESLint rule creation.
* Modify rule creation function to use TypeScript syntax.
* Update `eslint-plugin-test-rules/package.json` to reference `index.ts` instead of `index.js`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/w3bdesign/dfweb-v4/issues/576?shareId=XXXX-XXXX-XXXX-XXXX).
