+++ eslint-plugin-test-rules/package.json
@@ -2,7 +2,7 @@
   "name": "eslint-plugin-test-rules",
   "version": "1.0.0",
   "description": "Custom ESLint rules for testing",
-  "main": "index.js",
+  "main": "index.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
